backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.9
python-dotenv>=1.0
google-cloud-bigquery>=3.25
db-dtypes>=1.3
anthropic>=0.39
httpx>=0.27
tenacity>=9.0
